+++ src/ServiceProvider.php
@@ -21,16 +21,15 @@ namespace CloudCreativity\LaravelJsonApi;
 use CloudCreativity\JsonApi\Contracts\Exceptions\ExceptionParserInterface;
 use CloudCreativity\JsonApi\Contracts\Factories\FactoryInterface;
 use CloudCreativity\JsonApi\Contracts\Http\ApiInterface;
-use CloudCreativity\JsonApi\Contracts\Http\HttpServiceInterface;
 use CloudCreativity\JsonApi\Contracts\Http\Requests\RequestInterpreterInterface;
 use CloudCreativity\JsonApi\Contracts\Http\Responses\ResponseFactoryInterface;
 use CloudCreativity\JsonApi\Contracts\Repositories\ErrorRepositoryInterface;
 use CloudCreativity\JsonApi\Contracts\Store\StoreInterface;
+use CloudCreativity\JsonApi\Contracts\Validators\ValidatorErrorFactoryInterface;
 use CloudCreativity\JsonApi\Http\Responses\ResponseFactory;
 use CloudCreativity\LaravelJsonApi\Api\Repository;
 use CloudCreativity\LaravelJsonApi\Console\Commands;
 use CloudCreativity\LaravelJsonApi\Contracts\Document\LinkFactoryInterface;
-use CloudCreativity\LaravelJsonApi\Contracts\Validators\ValidatorErrorFactoryInterface;
 use CloudCreativity\LaravelJsonApi\Document\LinkFactory;
 use CloudCreativity\LaravelJsonApi\Exceptions\ExceptionParser;
 use CloudCreativity\LaravelJsonApi\Factories\Factory;
@@ -41,10 +40,12 @@ use CloudCreativity\LaravelJsonApi\Http\Requests\RequestInterpreter;
 use CloudCreativity\LaravelJsonApi\Http\Responses\Responses;
 use CloudCreativity\LaravelJsonApi\Services\JsonApiService;
 use CloudCreativity\LaravelJsonApi\Validators\ValidatorErrorFactory;
+use CloudCreativity\LaravelJsonApi\View\Renderer;
 use Illuminate\Contracts\Foundation\Application;
 use Illuminate\Contracts\Routing\ResponseFactory as ResponseFactoryContract;
 use Illuminate\Routing\Router;
 use Illuminate\Support\ServiceProvider as BaseServiceProvider;
+use Illuminate\View\Compilers\BladeCompiler;
 use Neomerx\JsonApi\Contracts\Document\DocumentFactoryInterface;
 use Neomerx\JsonApi\Contracts\Encoder\Handlers\HandlerFactoryInterface;
 use Neomerx\JsonApi\Contracts\Encoder\Parser\ParserFactoryInterface;
@@ -90,6 +91,7 @@ class ServiceProvider extends BaseServiceProvider
     ) {
         $this->bootMiddleware($router);
         $this->bootResponseMacro($responses);
+        $this->bootBladeDirectives();
     }
 
     /**
@@ -107,6 +109,7 @@ class ServiceProvider extends BaseServiceProvider
         $this->bindExceptionParser();
         $this->bindResponses();
         $this->bindLinkFactory();
+        $this->bindRenderer();
         $this->registerArtisanCommands();
         $this->mergePackageConfig();
     }
@@ -143,6 +146,17 @@ class ServiceProvider extends BaseServiceProvider
                 return app(ResponseFactoryInterface::class);
             });
         }
+    }
+
+    /**
+     * Register Blade directives.
+     */
+    protected function bootBladeDirectives()
+    {
+        /** @var BladeCompiler $compiler */
+        $compiler = $this->app->make(BladeCompiler::class);
+        $compiler->directive('jsonapi', Renderer::class . '::compileEncoder');
+        $compiler->directive('encode', Renderer::class . '::compileData');
     }
 
     /**
@@ -183,7 +197,6 @@ class ServiceProvider extends BaseServiceProvider
     protected function bindService()
     {
         $this->app->singleton(JsonApiService::class);
-        $this->app->alias(JsonApiService::class, HttpServiceInterface::class);
         $this->app->alias(JsonApiService::class, 'json-api.service');
     }
 
@@ -213,17 +226,10 @@ class ServiceProvider extends BaseServiceProvider
         });
 
         $this->app->bind(ValidatorErrorFactoryInterface::class, function (Application $app) {
-            /** @var Factory $factory */
-            $factory = $app->make(Factory::class);
             $errors = $app->make(ErrorRepositoryInterface::class);
 
             return new ValidatorErrorFactory($errors);
         });
-
-        $this->app->alias(
-            ValidatorErrorFactoryInterface::class,
-            \CloudCreativity\JsonApi\Contracts\Validators\ValidatorErrorFactoryInterface::class
-        );
     }
 
     /**
@@ -259,6 +265,14 @@ class ServiceProvider extends BaseServiceProvider
     {
         $this->app->singleton(LinkFactoryInterface::class, LinkFactory::class);
         $this->app->alias(LinkFactoryInterface::class, 'json-api.links');
+    }
+
+    /**
+     * Bind the view renderer into the service container.
+     */
+    protected function bindRenderer()
+    {
+        $this->app->singleton(Renderer::class);
     }
 
     /**